refactor(bigdecimal): atualiza WalletService e TransactionService para operações BigDecimal

+++ src/main/java/br/com/edras/picpaysimplificado/exception/wallet/InvalidAmountException.java
@@ -1,7 +1,9 @@
 package br.com.edras.picpaysimplificado.exception.wallet;
 
+import java.math.BigDecimal;
+
 public class InvalidAmountException extends RuntimeException {
-    public InvalidAmountException(Double amount) {
+    public InvalidAmountException(BigDecimal amount) {
         super("Valor inválido: R$" + amount);
     }
 }
